refac: [View] SignUp, LogIn, ToHash, GetToken
